Secure chat component edit message entity
